Updated uint64_t/int64_t to usize/isize

+++ src/command.c
@@ -11,7 +11,7 @@ void cmd_clear() { console_clear(); }
 
 void cmd_cd(int argc, const char** argv) {
     if (argc >= 2) {
-        int64_t status = set_current_working_directory(argv[1]);
+        isize status = set_current_working_directory(argv[1]);
         if (status < 0)
             printf("Error while opening %s: %s\n", argv[0], strerror(status));
     }
