+++ src/shell.h
@@ -1,9 +1,9 @@
 #ifndef __SHELL_H
 #define __SHELL_H
 
-#include <stdint.h>
+#include <los/types.h>
 
-uint64_t read_line(char** buffer);
+usize read_line(char** buffer);
 int parse_arguments(char* buffer, char*** argv);
 int run_internal_command(int argc, const char** argv);
